added blueprints and configurations
