+++ flaskblog/models.py
@@ -1,5 +1,6 @@
 from datetime import datetime
-from flaskblog import db, login_manager, app
+from flask import current_app
+from flaskblog import db, login_manager
 from flask_login import UserMixin
 from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
 
@@ -18,12 +19,12 @@ class User(db.Model, UserMixin):
     posts = db.relationship("Post", backref="author", lazy=True)
 
     def get_reset_token(self, expires_seconds=1800):
-        s = Serializer(app.config["SECRET_KEY"], expires_seconds)
+        s = Serializer(current_app.config["SECRET_KEY"], expires_seconds)
         return s.dumps({"user_id": self.id}).decode("utf-8")
 
     @staticmethod
     def verify_reset_token(token):
-        s = Serializer(app.config["SECRET_KEY"])
+        s = Serializer(current_app.config["SECRET_KEY"])
         try:
             user_id = s.loads(token)["user_id"]
         except:
