+++ flaskblog/templates/layout.html
@@ -18,24 +18,24 @@
         <header class="'site-header">
             <nav class="'navbar navbar-expand-md navbar-dark bg-steel fixed-top">
                 <div class="container">
-                    <a class="navbar-brand mr-4" href="{{ url_for('home') }}">Flask Blog</a>
+                    <a class="navbar-brand mr-4" href="{{ url_for('main.home') }}">Flask Blog</a>
                     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle" aria-controls="navbarToggle" aria-expanded="false" aria-label="Toggle navigation">
                         <span class="navbar-toggler-icon"></span>
                     </button>
                     <div class="collapse navbar-collapse" id="navbarToggle">
                         <div class="navbar-nav mr-auto">
-                            <a class="nav-item nav-link" href="{{ url_for('home') }}">Home</a>
-                            <a class="nav-item nav-link" href="{{ url_for('about') }}">About</a>
+                            <a class="nav-item nav-link" href="{{ url_for('main.home') }}">Home</a>
+                            <a class="nav-item nav-link" href="{{ url_for('main.about') }}">About</a>
                         </div>
                         <!-- Navbar Right Side -->
                         <div class="navbar-nav">
                             {% if current_user.is_authenticated %}
-                                <a class="nav-item nav-link" href="{{ url_for('new_post') }}">Create Post</a>
-                                <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-                                <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+                                <a class="nav-item nav-link" href="{{ url_for('posts.new_post') }}">Create Post</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.logout') }}">Logout</a>
                             {% else %}
-                                <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-                                <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.login') }}">Login</a>
+                                <a class="nav-item nav-link" href="{{ url_for('users.register') }}">Register</a>
                             {% endif %}
                         </div>
                     </div>
@@ -62,13 +62,13 @@
                         <p class="text-muted">Rotation Cities
                             <ul class="list-group list-group-item-action list-group-flush">
 
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">San Francisco</a>
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Seoul</a>
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Taipei</a>
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Hyderabad</a>
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Buenos Aires</a>
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">London</a>
-                                <a href="{{ url_for('new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Berlin</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">San Francisco</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Seoul</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Taipei</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Hyderabad</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Buenos Aires</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">London</a>
+                                <a href="{{ url_for('posts.new_post') }}" class="list-group-item list-group-item-action list-group-item-light">Berlin</a>
 
                             </ul>
                         </p>
