+++ flaskblog/templates/login.html
@@ -46,14 +46,14 @@
             <div class="form-group">
                 {{ form.submit(class="btn btn-outline-info") }}
                 <small class="text-muted ml-2">
-                    <a href="{{url_for('request_reset_password')}}">Forgot Password?</a>
+                    <a href="{{url_for('users.request_reset_password')}}">Forgot Password?</a>
                 </small>
             </div>
         </form>
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Need An Account? <a class="ml-2" href="{{url_for('register')}}">Sign up now</a>
+            Need An Account? <a class="ml-2" href="{{url_for('users.register')}}">Sign up now</a>
         </small>
     </div>
 {% endblock content %}
