+++ flaskblog/templates/register.html
@@ -78,7 +78,7 @@
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Already Have An Account? <a class="ml-2" href="{{url_for('login')}}">Log In</a>
+            Already Have An Account? <a class="ml-2" href="{{url_for('users.login')}}">Log In</a>
         </small>
     </div>
 {% endblock content %}
