+++ flaskblog/users/routes.py
@@ -1,0 +1,132 @@
+from flask import Blueprint, flash, redirect, url_for, render_template, request
+from flaskblog import db, bcrypt
+from flaskblog.users.forms import (
+    RegistrationForm,
+    LoginForm,
+    UpdateAccountForm,
+    RequestResetForm,
+    ResetPasswordForm,
+)
+from flaskblog.models import User, Post
+from flask_login import login_user, current_user, logout_user, login_required
+from flaskblog.users.utils import save_picture, send_reset_email
+
+users = Blueprint("users", __name__)
+
+
+@users.route("/register", methods=["GET", "POST"])
+def register():
+    if current_user.is_authenticated:
+        return redirect(url_for("main.home"))
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        hashed_pw = bcrypt.generate_password_hash(form.password.data).decode("utf-8")
+
+        user = User(
+            username=form.username.data, email=form.email.data, password=hashed_pw
+        )
+        db.session.add(user)
+        db.session.commit()
+
+        flash(f"Your account has been created! You are now able to log in", "success")
+        return redirect(url_for("users.login"))
+    return render_template("register.html", title="Register", form=form)
+
+
+@users.route("/login", methods=["GET", "POST"])
+def login():
+    if current_user.is_authenticated:
+        return redirect(url_for("main.home"))
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data).first()
+
+        if user and bcrypt.check_password_hash(user.password, form.password.data):
+            login_user(user, remember=form.remember.data)
+            next_page = request.args.get("next")
+            return redirect(next_page) if next_page else redirect(url_for("main.home"))
+        else:
+            flash("Login unsuccessful. Please check email and password", "danger")
+
+    return render_template("login.html", title="Login", form=form)
+
+
+@users.route("/logout")
+def logout():
+    logout_user()
+    return redirect(url_for("main.home"))
+
+
+@users.route("/account", methods=["GET", "POST"])
+@login_required
+def account():
+    image_file = url_for("static", filename="profile_pics/" + current_user.image_file)
+    form = UpdateAccountForm()
+
+    if form.validate_on_submit():
+        if form.picture.data:
+            picture_file = save_picture(form.picture.data)
+            current_user.image_file = picture_file
+            db.session.commit()
+
+        current_user.username = form.username.data
+        current_user.email = form.email.data
+        db.session.commit()
+        flash(f"Your account has been updated!", "success")
+        return redirect(url_for("users.account"))
+
+    elif request.method == "GET":
+        form.username.data = current_user.username
+        form.email.data = current_user.email
+
+    return render_template(
+        "account.html", title="Account", image_file=image_file, form=form
+    )
+
+
+@users.route("/user/<string:username>")
+def user_posts(username):
+    page = request.args.get("page", 1, type=int)
+    user = User.query.filter_by(username=username).first_or_404()
+    posts = (
+        Post.query.filter_by(author=user)
+        .order_by(Post.date_posted.desc())
+        .paginate(page=page, per_page=5)
+    )
+    return render_template("user_posts.html", posts=posts, user=user)
+
+
+@users.route("/reset_password", methods=["GET", "POST"])
+def request_reset_password():
+    if current_user.is_authenticated:
+        return redirect(url_for("main.home"))
+    form = RequestResetForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(email=form.email.data).first()
+        send_reset_email(user)
+        flash("Email to reset password sent", "info")
+        return redirect(url_for("users.login"))
+    return render_template("request_reset.html", form=form, title="Reset Password")
+
+
+@users.route("/reset_password/<token>", methods=["GET", "POST"])
+def reset_token(token):
+    if current_user.is_authenticated:
+        return redirect(url_for("main.home"))
+    user = User.verify_reset_token(token)
+    if not user:
+        flash("That is an invalid or expired token", "warning")
+        return redirect(url_for("users.request_reset_password"))
+
+    form = ResetPasswordForm()
+    if form.validate_on_submit():
+        hashed_pw = bcrypt.generate_password_hash(form.password.data).decode("utf-8")
+        user.password = hashed_pw
+        db.session.commit()
+
+        flash(
+            f"{user.username} Your password has been updated! You are now able to login",
+            "success",
+        )
+        return redirect(url_for("users.login"))
+    return render_template("reset_password.html", title="Reset Password", form=form)
