+++ run.py
@@ -1,4 +1,6 @@
-from flaskblog import app
+from flaskblog import create_app
 
-if __name__ == '__main__':
+app = create_app()
+
+if __name__ == "__main__":
     app.run(debug=True)
